Fix VitalsReading export error by creating shared types file
Create a centralized types.ts file for shared interfaces used across
multiple services (bluetoothService, healthPlatformService, vitalsSyncService).

This fixes the SyntaxError where healthPlatformService was unable to
import VitalsReading from bluetoothService. TypeScript interfaces are
type-only exports that get erased during compilation, causing runtime
module resolution errors.

Changes:
- Created web/src/services/types.ts with shared type definitions
- Updated bluetoothService.ts to import types from types.ts
- Updated healthPlatformService.ts to import types from types.ts
- Updated vitalsSyncService.ts to import types from types.ts
